requirements.txt
aiohttp==3.6.2
APScheduler==3.6.3
async-timeout==3.0.1
chardet==3.0.4
discord.py==1.3.1
emoji==0.5.4
multidict==4.7.4
pytz==2019.3
six==1.14.0
tzlocal==2.0.0
websockets==8.1
requests==2.22.0
beautifulsoup4==4.8.2
spotipy==2.12.0
regex==2020.1.8